refactor(product-builds): move pre-built packages tab to Package Analysis
Extract the Pre-Built Packages tab from Wheel Collections into its own
view component and add it as a new tab in Package Analysis, after the
Tracker tab. No changes to the server endpoint or composable.
